DCS-1187 imageController renamed to prisonerController
