+++ server/routes/index.ts
@@ -1,6 +1,6 @@
 import express, { RequestHandler, Router } from 'express'
 import ChoosePrisonerController from './choosePrisonerController'
-import ImageController from './imageController'
+import PrisonerController from './prisonerController'
 
 import asyncMiddleware from '../middleware/asyncMiddleware'
 import { Services } from '../services'
@@ -20,8 +20,8 @@ export default function routes(services: Services): Router {
   const temporaryAbsencesController = new TemporaryAbsencesController(services.temporaryAbsencesService)
   get('/confirm-arrival/return-from-temporary-absence', temporaryAbsencesController.view())
 
-  const imageController = new ImageController(services.incomingMovementsService)
-  get('/prisoner/:prisonNumber/image', imageController.getImage())
+  const prisonerController = new PrisonerController(services.incomingMovementsService)
+  get('/prisoner/:prisonNumber/image', prisonerController.getImage())
 
   return router
 }
