+++ server/routes/prisonerController.ts
@@ -3,7 +3,7 @@ import path from 'path'
 
 import IncomingMovementsService from '../services/incomingMovementsService'
 
-export default class ImageController {
+export default class PrisonerController {
   public constructor(private readonly incomingMovementsService: IncomingMovementsService) {}
 
   placeHolderImage = path.join(process.cwd(), '/assets/images/placeholder-image.png')
